+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Misplaced `Logo` position on mobile.
+
+### Changed
+
+- Changed `Logo` to be a functional component instead of a class.
+- Remove unused `VtexLogo` along with the props `color` and `showLabel` in `Logo`.
+- Replaced `withRuntime` HOC by `useRuntime` hook in `Logo`.
+
 ## [3.21.8] - 2019-04-03
 
 ### Fixed